+++ packages/server/utils/callGQL.ts
@@ -7,7 +7,7 @@ import {yoga} from '../yoga'
 // Only call this from the server that is running yoga! e.g. a webhook endpoint or similar
 export const callGQL = async <TData>(query: string, variables?: Record<string, any>) => {
   const authToken = new ServerAuthToken()
-  const dataLoader = getNewDataLoader()
+  const dataLoader = getNewDataLoader('callGQL')
   const {execute, parse} = yoga.getEnveloped()
   const result = await execute({
     document: parse(query),